socketio

+++ API.js
@@ -1,6 +1,6 @@
 import axios from "axios";
-const api_url = "https://twtr-clone-server.herokuapp.com"
-// const api_url = "http://localhost:5000"
+// const api_url = "https://twtr-clone-server.herokuapp.com"
+const api_url = "http://localhost:5000"
 
 const Server = {
     newUser: async (email, password, username, name) => {
@@ -34,36 +34,39 @@ const Server = {
 
 const entranceAlghorithm = async (context, localStorage, Router) => {
     const id = localStorage.getItem("id")
-    if (id && id != "undefined" && id != undefined) {
-        // console.log("ID FOUND: ", id);
-        Server.getUser({ id: id }).then(res => {
-            const user = res.data[0]
-            if (user) {
-                // console.log("USER FOUND: ", res.data[0]);
-                context.setUser(res.data[0])
-                context.setLoggedIn(true)
-            }
-            else {
+    if(!context.loggedIn){
+        if (id && id != "undefined" && id != undefined) {
+            // console.log("ID FOUND: ", id);
+            Server.getUser({ id: id }).then(res => {
+                const user = res.data[0]
+                if (user) {
+                    // console.log("USER FOUND: ", res.data[0]);
+                    context.setUser(res.data[0])
+                    context.setLoggedIn(true)
+                    context.socket.emit('logged-in', {id:user.id})
+                }
+                else {
+                    console.log("COULDNT FIND ANY USER WITH THE FOLLOWING ID:", id);
+                    context.setLoggedIn(false)
+                    context.setUser(null)
+                    localStorage.clear()
+                    Router.push("/login")
+                }
+            }).catch(err => {
                 console.log("COULDNT FIND ANY USER WITH THE FOLLOWING ID:", id);
                 context.setLoggedIn(false)
                 context.setUser(null)
                 localStorage.clear()
                 Router.push("/login")
-            }
-        }).catch(err => {
-            console.log("COULDNT FIND ANY USER WITH THE FOLLOWING ID:", id);
+            })
+        }
+        else {
+            console.log("COULDNT FIND ANY ID");
             context.setLoggedIn(false)
             context.setUser(null)
             localStorage.clear()
             Router.push("/login")
-        })
-    }
-    else {
-        console.log("COULDNT FIND ANY ID");
-        context.setLoggedIn(false)
-        context.setUser(null)
-        localStorage.clear()
-        Router.push("/login")
+        }
     }
 }
